turn off environment push in PushSim

+++ push_exp/PushWindowDraw.cpp
@@ -27,7 +27,7 @@ draw()
     mViewMatrix.translation() = b;
 
     auto ground = mEnv->GetGround();
-    float y = ground->getBodyNode(0)->getTransform().translation()[1] + dynamic_cast<const BoxShape*>(ground->getBodyNode(0)->getShapeNodesWith<dart::dynamics::VisualAspect>()[0]->getShape().get())->getSize()[1]*0.5;
+    double y = ground->getBodyNode(0)->getTransform().translation()[1] + dynamic_cast<const BoxShape*>(ground->getBodyNode(0)->getShapeNodesWith<dart::dynamics::VisualAspect>()[0]->getShape().get())->getSize()[1]*0.5;
 
     DrawGround(y);
     DrawTrajectory();
